feat: moved some parameters into e137_1
